Hibernate parcialmente aplicado
tendo dificuldade com o mapeamento de classe relacionadas

+++ src/IContrato/IEmpresaContacto.java
@@ -11,11 +11,11 @@ package IContrato;
  */
 public interface IEmpresaContacto {
     
-    int getEmpresaContactoId();
+    long getEmpresaContactoId();
     
     String getNome();
     
-    void setEmpresaContactoId(int empresaContactoId);
+    void setEmpresaContactoId(long empresaContactoId);
     
     void setNome(String nome);
 }
